Issue-133 (Multiple flames at same position - bomb movement/expl.)
When a flame dies it may reveal an item at the beginning of the step function.
If another flame at the same position is still alive it will replace the item on the board with a flame towards the end of the step function.
Between revealing the item on the board and replacing it with a flame the item will be visible to the bomb movement logic and will prevent the bomb from moving into the flame.

A similar issue when no item is revealed causes bombs to be able to move into a flame without exploding.

Just redraw all flames after all dead flames are handled.